requirements.txt
streamlit>=1.32
pandas>=2.2
numpy>=1.26
openpyxl>=3.1
plotly>=5.20
scikit-learn>=1.4
scipy>=1.12
anthropic>=0.40
python-dotenv>=1.0
